Split cost classes by type
